requirements.txt
tornado>=3.2
couchdb>=1.0.1
pyyaml>=3.10
